requirements.txt
ansible==2.9.0
appnope==0.1.0
asn1crypto==1.2.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.1.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
hvac==0.9.5
idna==2.8
imageio==2.8.0
imageio-ffmpeg==0.4.1
importlib-metadata==0.23
imutils==0.5.3
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kafka-python==1.4.7
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
moviepy==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
netaddr==0.7.19
notebook==6.0.2
numpy==1.17.3
opencv-python==4.1.1.26
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pbr==5.4.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.1.2
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
ptyprocess==0.6.0
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.5
pyrsistent==0.15.5
PySocks==1.7.1
pyspark==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.5.5
requests==2.22.0
rsa==4.0
ruamel-yaml==0.15.46
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
